"rebased to mysql"

+++ src/main/resources/db/migration/V1_1__Init_Data.sql
@@ -1,5 +1,5 @@
 create table users(
-                      id bigserial primary key,
+                      id serial primary key,
                       email varchar(255) not null ,
                       first_name varchar(255) not null ,
                       password text,
@@ -16,8 +16,8 @@ create table authorities(
 create unique index authorities_value_uindex on authorities(value);
 
 create table user_authorities(
-                                 user_id bigint not null ,
-                                 authority_id int not null ,
+                                 user_id bigint unsigned not null ,
+                                 authority_id bigint unsigned not null ,
                                  primary key (user_id, authority_id),
                                  constraint user_authorities_users_fk foreign key (user_id) references users(id),
                                  constraint user_authorities_authorities_fk foreign key (authority_id) references authorities(id)
